requirements.txt
# ============================
# 野生动物智能监测 - 第一阶段
# Python 3.8+ 
# 不需要GPU
# ============================

# 核心库
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# 图像/视频处理
opencv-python>=4.5.0
pillow>=8.3.0
imageio>=2.9.0
imageio-ffmpeg>=0.4.5  # 视频读取支持

# 音频处理
librosa>=0.9.0
soundfile>=0.10.0
audioread>=2.1.9

# 数据处理
torch>=1.10.0  # CPU版本即可
torchvision>=0.11.0  # CPU版本

# 可视化
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0

# 配置管理
pyyaml>=5.4.0
omegaconf>=2.1.0

# 进度条
tqdm>=4.62.0

# 文件操作
pathlib2>=2.3.6
glob2>=0.7


# Jupyter支持（可选）
jupyter>=1.0.0
notebook>=6.4.0
ipywidgets>=7.6.0

# 代码质量（可选）
black>=21.9b0
flake8>=3.9.0

# ??????
transformers>=4.20.0  # VideoMAE??
timm>=0.6.0           # ??????????